requirements.txt
# Django Backend Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0

# Authentication and authorization
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0

# For API documentation
drf-spectacular==0.26.5

# Development dependencies
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7

# Optional data handling (skip if installation fails on Windows)
# pandas==2.1.3
# numpy==1.25.2